parsInt en parsFloat

+++ js/main.js
@@ -1,8 +1,8 @@
 /**
  * Fichier de base pour application web HTML/CSS/JS
- * @author  Steve Fallet
+ * @author  Ursanne Salomon
  * @version 0.2
- * @since   2019-08-19
+ * @since   05.09.2023
  */
 'use strict'; // Demande un interprétation stricte du code
 
@@ -10,7 +10,7 @@ let a = 'Bonjour';
 
 // demande la température a l'utilisateur
 let celcius = prompt("Température en celcius : ");
-celcius = parseInt(celcius);
+celcius = parseFloat(celcius);
 
 // vérifie que la donnée saisie soit un nombre
 if(isNaN(celcius)) {
